prevent possible error

+++ scripts/update_apkid.sh
@@ -10,14 +10,16 @@ virtual_env=$(echo $VIRTUAL_ENV)
 apkid_dir="${virtual_env}/lib/python3.6/site-packages/apkid"
 rules_dir="${apkid_dir}/rules/"
 unamestr=`uname`
-git clone https://github.com/rednaga/APKiD.git 
-cd ./APKiD 
-python3 prep-release.py 
-cp apkid/rules/rules.yarc ${rules_dir}  
-cd ..
-if [[ "$unamestr" == 'Darwin' ]]; then
-  sed -i ' ' "s#RULES_DIR =.*#RULES_DIR =  \"$rules_dir\"#" ${apkid_dir}/rules.py 
- else
-  sed "s#RULES_DIR =.*#RULES_DIR =  \"$rules_dir\"#" ${apkid_dir}/rules.py
-fi 
+if [ -e "$virtual_env" ]; then
+    git clone https://github.com/rednaga/APKiD.git 
+    cd ./APKiD 
+    python3 prep-release.py 
+    cp apkid/rules/rules.yarc ${rules_dir}  
+    cd ..
+    if [[ "$unamestr" == 'Darwin' ]]; then
+        sed -i ' ' "s#RULES_DIR =.*#RULES_DIR =  \"$rules_dir\"#" ${apkid_dir}/rules.py 
+    else
+        sed "s#RULES_DIR =.*#RULES_DIR =  \"$rules_dir\"#" ${apkid_dir}/rules.py
+    fi 
 rm -fr ./APKiD
+fi
